Shared Input component
