+++ common/components/Input/index.ts
@@ -15,6 +15,5 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-export * from "./BinaryRadioButton";
-export * from "./Form.styles";
-export * from "./TabbedDisaggregations";
+export * from "./Input";
+export * from "./NotReportedIcon";
