+++ publisher/src/components/Reports/ReportSummaryPanel.tsx
@@ -16,6 +16,7 @@
 // =============================================================================
 
 import successIcon from "@justice-counts/common/assets/status-check-icon.png";
+import errorIcon from "@justice-counts/common/assets/status-error-icon.png";
 import {
   HEADER_BAR_HEIGHT,
   palette,
@@ -34,7 +35,6 @@ import {
   printDateRangeFromMonthYear,
   printElapsedDaysMonthsYearsSinceDate,
 } from "../../utils";
-import errorIcon from "../assets/status-error-icon.png";
 import { MetricSummarySectionTitle, Title } from "../Forms";
 import { REPORT_CAPITALIZED } from "../Global/constants";
 import { TeamMemberNameWithBadge } from "../primitives";
